add group relationship to project

+++ ambassador-gitlab-client/src/main/kotlin/com/roche/gitlab/api/project/model/SimpleProject.kt
@@ -49,7 +49,7 @@ open class SimpleProject(
 ) {
 
     @JsonProperty("namespace_id")
-    fun getNamespaceId(): Int? {
+    fun getNamespaceId(): Long? {
         return namespace?.id
     }
 }
